server/requirements.txt
fastapi==0.92.0
starlette==0.25.0
python-dotenv==1.0.0
uvicorn~=0.20.0